Moved wishToWishingWell function to wishingWells lib
The wishToWishingWell function, which helps in returning a JSON object with certain properties of an article, has been moved from the wishingWellGenerator script to the wishingWell library. This rearrangement enhances code modularity and organization.
